pyproject.toml
[tool.poetry]
name = "aerodisk-test-task"
version = "0.1.0"
description = ""
authors = ["RomaOkorosso <zatrutin.roma.work@gmail.com>"]
readme = "README.md"
packages = [{include = "aerodisk_test_task"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
pydantic = {extras = ["email"], version = "^1.10.7"}
black = "^23.3.0"
sqlalchemy = "^2.0.8"
passlib = {extras = ["cryptography"], version = "^1.7.4"}
alembic = "^1.10.2"
pyjwt = "^2.6.0"
python-dotenv = "^1.0.0"
jinja2 = "^3.1.2"
asyncpg = "^0.27.0"
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"
psycopg2 = "^2.9.6"
python-jose = "^3.3.0"
psutil = "^5.9.4"
wmi = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
